+++ taskmonitor/tests/test_models.py
@@ -59,6 +59,18 @@ class TestTaskLogQuerySet(TestCase):
         # then
         self.assertIsNone(result)
 
+    def test_should_return_stale_logs(self):
+        # given
+        current_dt = timezone.now()
+        TaskLogFactory(timestamp=current_dt)
+        log_2 = TaskLogFactory(timestamp=current_dt - dt.timedelta(hours=2))
+        log_3 = TaskLogFactory(timestamp=current_dt - dt.timedelta(hours=3))
+        # when
+        result = TaskLog.objects.filter_stale_logs(max_hours=1)
+        # then
+        pks = set(result.values_list("pk", flat=True))
+        self.assertSetEqual(pks, {log_2.pk, log_3.pk})
+
 
 class TestManagerCreateFromTask(TestCase):
     def test_should_create_from_succeeded_task(self):